moved alastriaSession and create its type
